% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark2003.R
\docType{data}
\name{mixed01}
\alias{mixed01}
\alias{mixed02}
\alias{mixed04}
\alias{mixed08}
\alias{mixed16}
\alias{mixed01_6000}
\alias{mixed02_6000}
\alias{mixed04_6000}
\alias{mixed08_6000}
\alias{mixed16_6000}
\alias{rural01}
\alias{rural02}
\alias{rural04}
\alias{rural08}
\alias{rural16}
\alias{rural01_6000}
\alias{rural02_6000}
\alias{rural04_6000}
\alias{rural08_6000}
\alias{rural16_6000}
\alias{rurmix01}
\alias{rurmix02}
\alias{rurmix04}
\alias{rurmix08}
\alias{rurmix16}
\alias{rurmix01_6000}
\alias{rurmix02_6000}
\alias{rurmix04_6000}
\alias{rurmix08_6000}
\alias{rurmix16_6000}
\alias{urban01}
\alias{urban02}
\alias{urban04}
\alias{urban08}
\alias{urban16}
\alias{urban01_6000}
\alias{urban02_6000}
\alias{urban04_6000}
\alias{urban08_6000}
\alias{urban16_6000}
\alias{urbmix01}
\alias{urbmix02}
\alias{urbmix04}
\alias{urbmix08}
\alias{urbmix16}
\alias{urbmix01_6000}
\alias{urbmix02_6000}
\alias{urbmix04_6000}
\alias{urbmix08_6000}
\alias{urbmix16_6000}
\alias{rururb01_6000}
\alias{rururb02_6000}
\alias{rururb04_6000}
\alias{rururb08_6000}
\alias{rururb16_6000}
\alias{rurmixurb01_6000}
\alias{rurmixurb02_6000}
\alias{rurmixurb04_6000}
\alias{rurmixurb08_6000}
\alias{rurmixurb16_6000}
\alias{rururb01}
\alias{rururb02}
\alias{rururb04}
\alias{rururb08}
\alias{rururb16}
\alias{rurmixurb01}
\alias{rurmixurb02}
\alias{rurmixurb04}
\alias{rurmixurb08}
\alias{rurmixurb16}
\title{Benchmark data from 2003}
\format{A matrix}
\description{
Benchmark data sets from Song and Kulldorff (2003).
Each data set is a 10000 x 245 matrix.  Each row
represents the data for a single simulated data
set.  The available data sets are
mixed01, mixed02, mixed04, mixed08, mixed16,
rural01, rural02, rural04, mixed08, rural16,
urban01, urban02, urban04, urban08, urban16,
urbmix01, urbmix02, urbmix04, urbmix08, urbmix16,
rurmix01, rurmix02, rurmix04, rurmix08, rurmix16,
rururb01, rururb02, rurub04, rururb08, rururb16,
rurmixurb01, rurmixurb02, rurmixurb04, rurmixurb08, rurmixurb16,
mixed01_6000, mixed02_6000, mixed04_6000, mixed08_6000, mixed16_6000,
rural01_6000, rural02_6000, rural04_6000, mixed08_6000, rural16_6000,
urban01_6000, urban02_6000, urban04_6000, urban08_6000, urban16_6000,
urbmix01_6000, urbmix02_6000, urbmix04_6000, urbmix08_6000, urbmix16_6000,
rurmix01_6000, rurmix02_6000, rurmix04_6000, rurmix08_6000, rurmix16_6000,
rururb01_6000, rururb02_6000, rururb04_6000, rururb08_6000, rururb16_6000,
rurmixurb01_6000, rurmixurb02_6000, rurmixurb04_6000, rurmixurb08_6000, rurmixurb16_6000.
}
\references{
Martin Kulldorff, Eric J. Feuer, Barry A.
  Miller, Laurence S. Freedman; Breast Cancer Clusters in
  the Northeast United States: A Geographic Analysis,
  American Journal of Epidemiology, Volume 146, Issue 2,
  15 July 1997, Pages 161-170,
  \url{https://doi.org/10.1093/oxfordjournals.aje.a009247}
  
Song, C., & Kulldorff, M. (2003). Power evaluation of disease clustering tests. International Journal of Health Geographics, 2, 9. http://doi.org/10.1186/1476-072X-2-9
}
\seealso{
neastdata
}
\keyword{data}
