% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark2003_hotspots.R
\docType{data}
\name{mixed01_hotspot}
\alias{mixed01_hotspot}
\alias{mixed02_hotspot}
\alias{mixed04_hotspot}
\alias{mixed08_hotspot}
\alias{mixed16_hotspot}
\alias{mixed01_6000_hotspot}
\alias{mixed02_6000_hotspot}
\alias{mixed04_6000_hotspot}
\alias{mixed08_6000_hotspot}
\alias{mixed16_6000_hotspot}
\alias{rural01_hotspot}
\alias{rural02_hotspot}
\alias{rural04_hotspot}
\alias{rural08_hotspot}
\alias{rural16_hotspot}
\alias{rural01_6000_hotspot}
\alias{rural02_6000_hotspot}
\alias{rural04_6000_hotspot}
\alias{rural08_6000_hotspot}
\alias{rural16_6000_hotspot}
\alias{rurmix01_hotspot}
\alias{rurmix02_hotspot}
\alias{rurmix04_hotspot}
\alias{rurmix08_hotspot}
\alias{rurmix16_hotspot}
\alias{rurmix01_6000_hotspot}
\alias{rurmix02_6000_hotspot}
\alias{rurmix04_6000_hotspot}
\alias{rurmix08_6000_hotspot}
\alias{rurmix16_6000_hotspot}
\alias{urban01_hotspot}
\alias{urban02_hotspot}
\alias{urban04_hotspot}
\alias{urban08_hotspot}
\alias{urban16_hotspot}
\alias{urban01_6000_hotspot}
\alias{urban02_6000_hotspot}
\alias{urban04_6000_hotspot}
\alias{urban08_6000_hotspot}
\alias{urban16_6000_hotspot}
\alias{urbmix01_hotspot}
\alias{urbmix02_hotspot}
\alias{urbmix04_hotspot}
\alias{urbmix08_hotspot}
\alias{urbmix16_hotspot}
\alias{urbmix01_6000_hotspot}
\alias{urbmix02_6000_hotspot}
\alias{urbmix04_6000_hotspot}
\alias{urbmix08_6000_hotspot}
\alias{urbmix16_6000_hotspot}
\alias{rururb01_6000_hotspot}
\alias{rururb02_6000_hotspot}
\alias{rururb04_6000_hotspot}
\alias{rururb08_6000_hotspot}
\alias{rururb16_6000_hotspot}
\alias{rurmixurb01_6000_hotspot}
\alias{rurmixurb02_6000_hotspot}
\alias{rurmixurb04_6000_hotspot}
\alias{rurmixurb08_6000_hotspot}
\alias{rurmixurb16_6000_hotspot}
\alias{rururb01_hotspot}
\alias{rururb02_hotspot}
\alias{rururb04_hotspot}
\alias{rururb08_hotspot}
\alias{rururb16_hotspot}
\alias{rurmixurb01_hotspot}
\alias{rurmixurb02_hotspot}
\alias{rurmixurb04_hotspot}
\alias{rurmixurb08_hotspot}
\alias{rurmixurb16_hotspot}
\title{Benchmark data from 2003}
\format{A vector}
\description{
The hotspot region indices for each benchmark
data set from Song and Kulldorff (2003).
Each data set is a vector of hotspot indices.  
The available data sets are
mixed01_hotspot, mixed02_hotspot, mixed04_hotspot, mixed08_hotspot, mixed16_hotspot,
rural01_hotspot, rural02_hotspot, rural04_hotspot, mixed08_hotspot, rural16_hotspot,
urban01_hotspot, urban02_hotspot, urban04_hotspot, urban08_hotspot, urban16_hotspot,
urbmix01_hotspot, urbmix02_hotspot, urbmix04_hotspot, urbmix08_hotspot, urbmix16_hotspot,
rurmix01_hotspot, rurmix02_hotspot, rurmix04_hotspot, rurmix08_hotspot, rurmix16_hotspot,
rururb01_hotspot, rururb_02_hotspot, rururb04_hotspot, rururb08_hotspot, rururb16_hotspot,
rurmixurb01_hotspot, rurmixurb02_hotspot, rurmixurb04_hotspot, rurmixurb08_hotspot, rurmixurb16_hotspot,
mixed01_6000_hotspot, mixed02_6000_hotspot, mixed04_6000_hotspot, mixed08_6000_hotspot, mixed16_6000_hotspot,
rural01_6000_hotspot, rural02_6000_hotspot, rural04_6000_hotspot, mixed08_6000_hotspot, rural16_6000_hotspot,
urban01_6000_hotspot, urban02_6000_hotspot, urban04_6000_hotspot, urban08_6000_hotspot, urban16_6000_hotspot,
urbmix01_6000_hotspot, urbmix02_6000_hotspot, urbmix04_6000_hotspot, urbmix08_6000_hotspot, urbmix16_6000_hotspot,
rurmix01_6000_hotspot, rurmix02_6000_hotspot, rurmix04_6000_hotspot, rurmix08_6000_hotspot, rurmix16_6000_hotspot,
rururb01_6000_hotspot, rururb02_6000_hotspot, rururb04_6000_hotspot, rururb08_6000_hotspot, rururb16_6000_hotspot,
rurmixurb01_6000_hotspot, rurmixurb_02_6000_hotspot, rurmixurb04_6000_hotspot, rurmixurb08_6000_hotspot, rurmixurb16_6000_hotspot.
}
\references{
Martin Kulldorff, Eric J. Feuer, Barry A.
  Miller, Laurence S. Freedman; Breast Cancer Clusters in
  the Northeast United States: A Geographic Analysis,
  American Journal of Epidemiology, Volume 146, Issue 2,
  15 July 1997, Pages 16_hotspot1-170,
  \url{https://doi.org/10.1093/oxfordjournals.aje.a009247}
  
Song, C., & Kulldorff, M. (2003). Power evaluation of disease clustering tests. International Journal of Health Geographics, 2, 9. http://doi.org/10.1186/1476-072X-2-9
}
\seealso{
neastdata
}
\keyword{data}
