% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark2006_hotpots.R
\docType{data}
\name{a_hotspot}
\alias{a_hotspot}
\alias{b_hotspot}
\alias{c_hotspot}
\alias{d_hotspot}
\alias{e_hotspot}
\alias{f_hotspot}
\alias{g_hotspot}
\alias{h_hotspot}
\alias{i_hotspot}
\alias{j_hotspot}
\alias{k_hotspot}
\title{Hotspots for benchmark2006 data}
\format{A matrix}
\description{
The hotspot region indices for each benchmark
data set 
from Duczmal et al. (2006).  Each data set is a 
vector of hotspot indices.  Available data sets are
a_hotspot, b_hotspot, c_hotspot, d_hotspot, 
e_hotspot, f_hotspot, g_hotspot, h_hotspot, 
i_hotspot, j_hotspot, k_hotspot.
}
\references{
Duczmal, L., Kulldorff, M., & Huang, L. (2006). Evaluation of Spatial Scan Statistics for Irregularly Shaped Clusters. Journal of Computational and Graphical Statistics, 15(2), 428-442. Retrieved from http://www.jstor.org/stable/27594187
}
\seealso{
neastdata
}
\keyword{data}
