% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark2020.R
\docType{data}
\name{irural02}
\alias{irural02}
\alias{iclusters}
\alias{irural03}
\alias{irural04}
\alias{irural05}
\alias{irural06}
\alias{irural07}
\alias{irural08}
\alias{irural09}
\alias{irural10}
\alias{irural11}
\alias{irural12}
\alias{irural13}
\alias{irural14}
\alias{irural15}
\alias{irural16}
\alias{imixed02}
\alias{imixed03}
\alias{imixed04}
\alias{imixed05}
\alias{imixed06}
\alias{imixed07}
\alias{imixed08}
\alias{imixed09}
\alias{imixed10}
\alias{imixed11}
\alias{imixed12}
\alias{imixed13}
\alias{imixed14}
\alias{imixed15}
\alias{imixed16}
\alias{iurban02}
\alias{iurban03}
\alias{iurban04}
\alias{iurban05}
\alias{iurban06}
\alias{iurban07}
\alias{iurban08}
\alias{iurban09}
\alias{iurban10}
\alias{iurban11}
\alias{iurban12}
\alias{iurban13}
\alias{iurban14}
\alias{iurban15}
\alias{iurban16}
\title{Irregularly-shaped connected benchmark data from 2020}
\format{
A matrix
}
\description{
Benchmark data sets with irregularly shaped clusters 
from French et al. (2023).  Each data set is 10000 x
245 matrix.  Each row provides the number of cases
in each region for a particular simulated data set.
Available data sets are \code{irural02}, \code{irural03}, ...,
\code{irural16}, \code{imixed02}, \code{imixed03}, ...
\code{imixed16}, \code{iurban02}, \code{iurban03}, ...
\code{iurban16}
}
\references{
Joshua P. French, Mohammad Meysami, Lauren M.
Hall, Nicholas E. Weaver, Minh C. Nguyen & Lee Panter
(2022) A comparison of spatial scan methods for cluster
detection, Journal of Statistical Computation and
Simulation, 92:16, 3343-3372.
}
\seealso{
neastdata
}
\keyword{data}
