% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark2020_hotpots.R
\docType{data}
\name{irural02_hotspot}
\alias{irural02_hotspot}
\alias{irural03_hotspot}
\alias{irural04_hotspot}
\alias{irural05_hotspot}
\alias{irural06_hotspot}
\alias{irural07_hotspot}
\alias{irural08_hotspot}
\alias{irural09_hotspot}
\alias{irural10_hotspot}
\alias{irural11_hotspot}
\alias{irural12_hotspot}
\alias{irural13_hotspot}
\alias{irural14_hotspot}
\alias{irural15_hotspot}
\alias{irural16_hotspot}
\alias{imixed02_hotspot}
\alias{imixed03_hotspot}
\alias{imixed04_hotspot}
\alias{imixed05_hotspot}
\alias{imixed06_hotspot}
\alias{imixed07_hotspot}
\alias{imixed08_hotspot}
\alias{imixed09_hotspot}
\alias{imixed10_hotspot}
\alias{imixed11_hotspot}
\alias{imixed12_hotspot}
\alias{imixed13_hotspot}
\alias{imixed14_hotspot}
\alias{imixed15_hotspot}
\alias{imixed16_hotspot}
\alias{iurban02_hotspot}
\alias{iurban03_hotspot}
\alias{iurban04_hotspot}
\alias{iurban05_hotspot}
\alias{iurban06_hotspot}
\alias{iurban07_hotspot}
\alias{iurban08_hotspot}
\alias{iurban09_hotspot}
\alias{iurban10_hotspot}
\alias{iurban11_hotspot}
\alias{iurban12_hotspot}
\alias{iurban13_hotspot}
\alias{iurban14_hotspot}
\alias{iurban15_hotspot}
\alias{iurban16_hotspot}
\title{Hotspots for benchmark2020 data}
\format{
A matrix
}
\description{
The hotspot region indices for each benchmark
data set 
from French et al. (2020+). These are randomly selected
subsets of connected regions of size 2, 3, ..., 16 from 
the 50 nearest neighbors of \code{rural01_hotspot},
\code{mixed01_hotspot}, \code{urban01_hotspot}. Each hotspot
must include the starting centroid.
}
\seealso{
neastdata
}
\keyword{data}
