% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{cpower}
\alias{cpower}
\title{Compute power of test}
\usage{
cpower(tnull, tdata, hotspot, type = "basic", alpha = c(0.05, 0.01))
}
\arguments{
\item{tnull}{The set of null test statistics}

\item{tdata}{The list of maximum test statistics (\code{tmax}) and most likely cluster (\code{mlc}) for each simulated data set.}

\item{hotspot}{A vector containing the hotspot indices for the current data set.}

\item{type}{The type of power to compute.  \code{"basic"} simply determines whether a cluster is detected at the provided \code{alpha} leves.  \code{type = "touch"} computes
the proportion of tests where the significant clusters intersect the true cluster. \code{"overlap"} computers the average proportion of the hotspot detected across all simulations.}

\item{alpha}{The type I error rate.  Default is \code{c(0.05, 0.01)}.}
}
\value{
A vector of power calculations.
}
\description{
Compute the empirical power of a method from the null
test statistics, the results from the observed data,
and the true hotspot locations.  The types of power
include \code{basic}, which is simply the proportion of
observed data sets that a significant result was detected,
\code{intersect}, which is the proportion of times a
significant most likely cluster intersect the true hotspot,
\code{overlap}, which is the average of the proportion of  
the hotspot detected in each simulated data set, and 
\code{contain}, which is the proportion of times a significant
most likely cluster entirely contained the true hotspot.
}
\examples{
tnull = 1:99
tdata = list(list(tmax = 96, mlc = c(50, 51)),
             list(tmax = 101, mlc = c(48, 57)))
cpower(tnull, tdata, 50, type = "basic")
cpower(tnull, tdata, 50, type = "intersect")
cpower(tnull, tdata, 50, type = "overlap")
cpower(tnull, tdata, 50, type = "contain")
}
