% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakenull.R
\docType{data}
\name{fakedata1}
\alias{fakedata1}
\alias{fakedata2}
\title{Fake data sets}
\format{
A 10 x 245 matrix.
}
\description{
There are two "fake" data sets.  Each row
represents a different simulated data set under the
constant risk hypothesis.  The fake data sets are
\code{fakedata1} and \code{fakedata2}
}
