% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neast_sf.R
\docType{data}
\name{neast_sf}
\alias{neast_sf}
\title{Breast cancer mortality in the Northeastern United States}
\description{
An \code{\link[sf]{sf}} object containing data related to
breast cancer mortality in the Northeastern United
States.  The data include information from 245 counties
(or similar) and polygon information converted to
coordinate system EPSG 2263. The following variables are
included in the object:
\itemize{
\item \code{id}: The name of each county.
\item \code{population}: The number of residents in the county
based on 1990 U.S. census results.
\item \code{easting}: The easting coordinate of a centroid associated with each county.
\item \code{northing}: The northing coordinate of a centroid associated with each county.
\item \code{cases}: The number of breast cancer
mortality cases between 1988-1992.
}
}
\examples{
if (require(sf)) {
data(neast_sf)
plot(st_geometry(neast_sf))
plot(neast_sf["cases"])
}
}
\references{
Martin Kulldorff, Eric J. Feuer, Barry A.
  Miller, Laurence S. Freedman; Breast Cancer Clusters in
  the Northeast United States: A Geographic Analysis,
  American Journal of Epidemiology, Volume 146, Issue 2,
  15 July 1997, Pages 161–170,
  \url{https://doi.org/10.1093/oxfordjournals.aje.a009247}
}
\keyword{data}
