% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neastdata.R
\docType{data}
\name{neastdata}
\alias{neastdata}
\title{Breast cancer mortality in the Northeastern United States}
\description{
Data related to breast cancer mortality in the
Northeastern United States.  The data include information
from 245 counties (or similar). The name of each county
is included, as well as the number of breast cancer
mortality cases between 1988-1992.  The population values
are based on 1990 U.S. census results.  A representative
centroid of each region is proved in Easting and Northing
coordinates.
}
\references{
Martin Kulldorff, Eric J. Feuer, Barry A.
  Miller, Laurence S. Freedman; Breast Cancer Clusters in
  the Northeast United States: A Geographic Analysis,
  American Journal of Epidemiology, Volume 146, Issue 2,
  15 July 1997, Pages 161–170,
  \url{https://doi.org/10.1093/oxfordjournals.aje.a009247}
}
\keyword{data}
