% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neastpoly.R
\docType{data}
\name{neastpoly}
\alias{neastpoly}
\title{\code{SpatialPolygons} object for \code{neastdata}}
\description{
This loads a \code{SpatialPolygons} providing the borders
for \code{neastdata}.
}
\examples{
if (require(sp)) {
data(neastpoly)
plot(neastpoly)
}
}
\references{
Martin Kulldorff, Eric J. Feuer, Barry A.
  Miller, Laurence S. Freedman; Breast Cancer Clusters in
  the Northeast United States: A Geographic Analysis,
  American Journal of Epidemiology, Volume 146, Issue 2,
  15 July 1997, Pages 161–170,
  \url{https://doi.org/10.1093/oxfordjournals.aje.a009247}
}
\seealso{
neastdata
}
\keyword{data}
