% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv.R
\name{npv}
\alias{npv}
\title{Compute negative predictive value}
\usage{
npv(tnull, tdata, hotspot, pop, alpha = c(0.05, 0.01))
}
\arguments{
\item{tnull}{The set of null test statistics}

\item{tdata}{The list of maximum test statistics
(\code{tmax}) and most likely cluster (\code{mlc}) for
each simulated data set.}

\item{hotspot}{A vector containing the hotspot indices
for the current data set.}

\item{pop}{A vector with the populations associated with each region.}

\item{alpha}{The type I error rate.  Default is
\code{c(0.05, 0.01)}.}
}
\value{
A vector of npv values.
}
\description{
Compute the empirical negative predictive value (npv)
rate of a method from a series of tests.  In this
context, the npv is the average proportion of the
population that lies outside the most
likely cluster that intersects the true null regoin.  
The function requires the null test
statistics, the results from the observed data sets
(i.e., the maximum test statistic and most likely cluster
from each data set), the true hotspot locations, and the
vector of population sizes for each region.  See Details.
}
\details{
In this context, the npv is the proportion of the
population lying outside the most likely cluster that
intersects the true null regoins, averaged over all tests.
}
\examples{
tnull = 1:99
tdata = list(list(tmax = 96, mlc = c(50, 51)),
             list(tmax = 101, mlc = c(48, 57)))
npv(tnull, tdata, 50, pop = rep(10, 100))
}
