% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{Compute power of test}
\usage{
power(tnull, tdata, hotspot, type = "basic", alpha = c(0.05, 0.01))
}
\arguments{
\item{tnull}{The set of null test statistics}

\item{tdata}{The list of maximum test statistics (\code{tmax}) and most likely cluster (\code{mlc}) for each simulated data set.}

\item{hotspot}{A vector containing the hotspot indices for the current data set.}

\item{type}{The type of power to compute.  \code{"basic"} simply determines whether a cluster is detected at the provided \code{alpha} leves.  \code{type = "touch"} computes
the proportion of tests where the significant clusters intersect the true cluster. \code{"overlap"} computers the average proportion of the hotspot detected across all simulations.}

\item{alpha}{The type I error rate.  Default is \code{c(0.05, 0.01)}.}
}
\value{
A vector of power calculations.
}
\description{
Compute power of test
}
\examples{
tnull = 1:99
tdata = list(list(tmax = 96, mlc = c(50, 51)),
             list(tmax = 101, mlc = c(48, 57)))
power(tnull, tdata, 50, type = "basic")
power(tnull, tdata, 50, type = "touch")
power(tnull, tdata, 50, type = "overlap")
}
