% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv.R
\name{ppv}
\alias{ppv}
\alias{precision}
\title{Compute positive predictive value}
\usage{
ppv(tnull, tdata, hotspot, pop, alpha = c(0.05, 0.01))

precision(tnull, tdata, hotspot, pop, alpha = c(0.05, 0.01))
}
\arguments{
\item{tnull}{The set of null test statistics}

\item{tdata}{The list of maximum test statistics
(\code{tmax}) and most likely cluster (\code{mlc}) for
each simulated data set.}

\item{hotspot}{A vector containing the hotspot indices
for the current data set.}

\item{pop}{A vector with the populations associated with each region.}

\item{alpha}{The type I error rate.  Default is
\code{c(0.05, 0.01)}.}
}
\value{
A vector of ppvs.
}
\description{
Compute the empirical positive predictive value (ppv) of a
method from a series of tests.  In this context, the
ppv is the average proportion of the true hostpot that
intersects the most likely cluster.
The function requires the null test statistics,
the results from the observed data sets (i.e., the
maximum test statistic and most likely cluster from each
data set), the true hotspot locations, and the vector of
population sizes for each region.
}
\examples{
tnull = 1:99
tdata = list(list(tmax = 96, mlc = c(50, 51)),
             list(tmax = 101, mlc = c(48, 57)))
ppv(tnull, tdata, 50, pop = rep(1, 100))
}
