% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\alias{recall}
\alias{tpr}
\title{Compute sensitivity/recall/true positive rate}
\usage{
sensitivity(tnull, tdata, hotspot, pop, alpha = c(0.05, 0.01))

recall(tnull, tdata, hotspot, pop, alpha = c(0.05, 0.01))

tpr(tnull, tdata, hotspot, pop, alpha = c(0.05, 0.01))
}
\arguments{
\item{tnull}{The set of null test statistics}

\item{tdata}{The list of maximum test statistics
(\code{tmax}) and most likely cluster (\code{mlc}) for
each simulated data set.}

\item{hotspot}{A vector containing the hotspot indices
for the current data set.}

\item{pop}{A vector with the populations associated with each region.}

\item{alpha}{The type I error rate.  Default is
\code{c(0.05, 0.01)}.}
}
\value{
A vector of sensitivity values.
}
\description{
Compute the empirical sensitivity/recall/true positive
rate of a method from a series of tests.  In this
context, the sensitivity is the average proportion of the
population of the true hotspot that lies within the most
likely cluster.  The function requires the null test
statistics, the results from the observed data sets
(i.e., the maximum test statistic and most likely cluster
from each data set), the true hotspot locations, and the
vector of population sizes for each region.  See Details.
}
\details{
In this context, the sensitivity is the proportion of the
true hotspot population lying within the most likely
cluster, averaged over all tests.  If the \code{pop}
vector is a vector of 1s, then the sensitivity will be
the average of the proportion of the most likely cluster
regions intersecting the true hotspot, divided by the
number of regions in the true hotspot.
}
\examples{
tnull = 1:99
tdata = list(list(tmax = 96, mlc = c(50, 51)),
             list(tmax = 101, mlc = c(48, 57)))
sensitivity(tnull, tdata, 50, pop = rep(1, 100))
}
