% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{specificity}
\alias{specificity}
\alias{selectivity}
\alias{tnr}
\title{Compute specificity/selectivity/true negative rate}
\usage{
specificity(tnull, tdata, hotspot, pop, alpha = c(0.05, 0.01))

selectivity(tnull, tdata, hotspot, pop, alpha = c(0.05, 0.01))

tnr(tnull, tdata, hotspot, pop, alpha = c(0.05, 0.01))
}
\arguments{
\item{tnull}{The set of null test statistics}

\item{tdata}{The list of maximum test statistics
(\code{tmax}) and most likely cluster (\code{mlc}) for
each simulated data set.}

\item{hotspot}{A vector containing the hotspot indices
for the current data set.}

\item{pop}{A vector with the populations associated with each region.}

\item{alpha}{The type I error rate.  Default is
\code{c(0.05, 0.01)}.}
}
\value{
A vector of specificity values.
}
\description{
Compute the empirical specificity/selectivity/true negative
rate of a method from a series of tests.  In this
context, the specificity is the average proportion of the
population of the true null region that lies outside the most
likely cluster.  The function requires the null test
statistics, the results from the observed data sets
(i.e., the maximum test statistic and most likely cluster
from each data set), the true hotspot locations, and the
vector of population sizes for each region.  See Details.
}
\details{
In this context, the specificity is the proportion of the
true null region population lying outside the most likely
cluster, averaged over all tests.
}
\examples{
tnull = 1:99
tdata = list(list(tmax = 96, mlc = c(50, 51)),
             list(tmax = 101, mlc = c(48, 57)))
specificity(tnull, tdata, 50, pop = rep(10, 100))
}
