% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.power.R
\name{benchmark.power}
\alias{benchmark.power}
\title{Benchmark power of test}
\usage{
benchmark.power(test.name, data.name, alpha = c(0.05, 0.01))
}
\arguments{
\item{test.name}{The name of the test being applied.}

\item{data.name}{The name(s) of the data sets to be tested.}

\item{alpha}{The type I error rate.  Default is \code{c(0.05, 0.01)}.}
}
\value{
A data.frame
}
\description{
Benchmark power of test
}
\examples{
\dontrun{
benchmark.power("scan_stat", 
                data.name = c("mixed01", "mixed01_6000"))
}
}
