% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neastdata.R
\docType{data}
\name{neastdata}
\alias{neastdata}
\title{Breast cancer mortality in the Northeastern United States}
\description{
Data related to breast cancer mortality in the
Northeastern United States.  The data include information
from 245 counties (or similar). The following variables
are included in the data frame:
\itemize{
\item \code{id}: A name-based id for each county.
\item \code{cases}: The number of breast cancer
mortality cases between 1988-1992.
\item \code{population}: The number of residents in the county
based on 1990 U.S. census results.
\item \code{x}: An x coordinate of a centroid associated
with each county provided by Kulldorff et al. (2003). See
Details.
\item \code{y}: A y coordinate of a centroid associated
with each county provided by Kulldorff et al. (2003). See
Details.
\item \code{easting}: Identical to \code{x}. Retained for
reverse compatibility.
\item \code{northing}: Identical to \code{y}. Retained for
reverse compatibility.
}
}
\details{
The \code{x} and \code{y} coordinates define centroids
associated with each county. The coordinates were
provided by Kulldorf et al. (2003). The coordinate system
of these coordinates is unknown.
}
\references{
Martin Kulldorff, Eric J. Feuer, Barry A.
  Miller, Laurence S. Freedman; Breast Cancer Clusters in
  the Northeast United States: A Geographic Analysis,
  American Journal of Epidemiology, Volume 146, Issue 2,
  15 July 1997, Pages 161–170,
  \url{https://doi.org/10.1093/oxfordjournals.aje.a009247}
}
\keyword{data}
